+++ src/marqo/base_model.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
 
 class StrictBaseModel(BaseModel):
