+++ src/marqo/core/models/marqo_index_health.py
@@ -2,7 +2,7 @@ from enum import Enum
 from types import DynamicClassAttribute
 from typing import Optional
 
-from marqo.core.models.strict_base_model import StrictBaseModel
+from marqo.base_model import StrictBaseModel
 
 
 class HealthStatus(Enum):
