+++ src/marqo/core/models/marqo_index_stats.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from marqo.core.models.strict_base_model import StrictBaseModel
+from marqo.base_model import StrictBaseModel
 
 
 class VespaStats(StrictBaseModel):
